Fixed some bugs, and merged some test files
