+++ pkg/lib/inventory/container/container_test.go
