+++ pkg/lib/inventory/model/client_test.go
